[fix]:fix the bug of second run 'node ./testCaseFactory/createCase.js'
